Move allocation logic outside of executor.
