+++ csrc/multidevice/executor.cpp
@@ -7,6 +7,7 @@
 // clang-format on
 #include <ATen/cuda/CUDAContext.h>
 #include <device_lower/utils.h>
+#include <fusion_executor/allocations.h>
 #include <fusion_segmenter.h>
 #include <host_ir/container.h>
 #include <host_ir/host_ir.h>
