+++ tests/cpp/utils.h
@@ -12,6 +12,7 @@
 #include <device_lower/lower2device.h>
 #include <device_lower/pass/magic_zero.h>
 #include <expr_evaluator.h>
+#include <fusion_executor/allocations.h>
 #include <fusion_executor/executor.h>
 #include <id_model/id_model.h>
 #include <ir/all_nodes.h>
